versión 8.0
tuned axe and superaxe values
super axe shot in all situations
enemy defeated state
enemy receive hit
tuned weapons
dino receives hit, close eye, die...
microjump, for steping enemies
player reacts at attacks
audio module
music
axe sounds
axe last moment added (40 milliseconds on screen before disappear)

+++ CavemanNinja/CavemanNinja/Collider.cpp
@@ -23,8 +23,8 @@ void Collider::SetPos(int x, int y, bool can_flip)
 		}
 	}
 
-	rect.x = (x * SCREEN_SIZE) + (float)temp_offset.x;
-	rect.y = (y * SCREEN_SIZE) + (float)temp_offset.y;
+	rect.x = (x * SCREEN_SIZE) + (int)temp_offset.x;
+	rect.y = (y * SCREEN_SIZE) + (int)temp_offset.y;
 }
 
 bool Collider::IsColliding(Collider* collider) const
